🚧 re-adicionada interação de filtros com botões do touch

+++ Application/DSP/Inc/audio_user_dsp.h
@@ -5,7 +5,7 @@
 #include "stdbool.h"
 
 void AudioUserDsp_ApplyFilterToSamples(uint8_t* dataPointer, uint32_t dataLength, int16_t (*leftChannelFilter)(int16_t, uint32_t), int16_t (*rightChannelFilter)(int16_t, uint32_t));
-int16_t AudioUserDsp_ChangeAmplitude(int16_t sample, uint16_t amplitude);
+int16_t AudioUserDsp_ChangeAmplitude(int16_t sample, uint32_t amplitude);
 int16_t AudioUserDsp_LowPassFilter(int16_t sample, uint32_t iteration);
 
 #endif // __AUDIO_USER_DSP_H__
